Creating Calendars.md
This may become broken down into smaller files, as a lot of research was conducted, and more thorough synthesis is needed.
I remain somewhat unconvinced of the historically religious nature of the 28-day 13-month calendar, though it does tend to keep closer to a true Luna-Solar calendar, and does so accurately within scientific and historical contexts.  Would like to find more references to the Essene Calendar, as this seems to be the closest to the "Original" Judaic Christian calendars.

Signed-off-by: Meta <59613197+metagrapher@users.noreply.github.com>